+++ app/auth/auth_model.py
@@ -1,0 +1,8 @@
+class User:
+
+    def __init__(self, user_id, name, username, password, admin_status):
+        self.user_id = user_id
+        self.name = name
+        self.username = username
+        self.password = password
+        self.admin_status = admin_status
