+++ app/menu/menu_model.py
@@ -1,0 +1,7 @@
+class Menu:
+    '''  Defines a Menu class'''
+    def __init__(self, item_id, item_name, price):
+        ''' Initializes the order object'''
+        self.item_id = item_id
+        self.item_name = item_name
+        self.price = price
