+++ run.py
@@ -1,7 +1,7 @@
 import os
 import psycopg2
 from app import create_app
-from app.database import Database
+from app.database.database import *
 
 app = create_app('DEFAULT')
 db = Database('postgresql://postgres:k0779211758aj@localhost:5432/order_db')
